minor refactoring in the code
