+++ components/BuildingDetailsPanel.jsx
@@ -132,7 +132,7 @@ const styles = StyleSheet.create({
         backgroundColor: theme.colors.white,
         borderRadius: 20,
         padding: 5,
-        shadowColor: "#000",
+        shadowColor: theme.colors.dark,
         shadowOffset: {width: 0, height: 3},
         shadowOpacity: 0.2,
         shadowRadius: 5,
@@ -194,7 +194,7 @@ const styles = StyleSheet.create({
         borderRadius: 15,
         alignItems: "center",
         justifyContent: "center",
-        shadowColor: "#000",
+        shadowColor: theme.colors.dark,
         shadowOffset: { width: 0, height: 2 },
         shadowOpacity: 0.15,
         shadowRadius: 5,
